Правит отображение звания пользователя
